+++ crates/bedit-cloudburst/src/files.rs
@@ -17,8 +17,12 @@ use serde::{Deserialize, Serialize};
 use serde_with::skip_serializing_none;
 use std::{
     collections::{btree_map, BTreeMap, VecDeque},
+    fs::File,
+    iter::FusedIterator,
+    marker::PhantomData,
     num::NonZeroU64,
     path::{Path, PathBuf},
+    rc::Rc,
 };
 
 #[cfg(debug_assertions)]
@@ -28,7 +32,7 @@ use log::{debug, error};
 
 /// Files shared by the torrent if multiple as per meta version 1.
 #[skip_serializing_none]
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 pub struct SharedFiles {
     /// File attribute such as whether the file is executable or hidden.
@@ -54,7 +58,7 @@ pub struct SharedFiles {
 /// V2 torrents use a different encoding scheme for files. Files and directories are stored as a tree where
 /// the leaf nodes describe files.
 #[skip_serializing_none]
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 pub struct FileTreeInfo {
     /// File attribute such as whether a file is executable or hidden.
@@ -83,27 +87,27 @@ pub struct FileTreeInfo {
 ///
 /// # Ok::<(), Error>(())
 /// ```
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
 #[serde(transparent)]
 pub struct FileTreeEntry(
     #[serde(with = "either::serde_untagged")] pub Either<FileTreeInfo, FileTree>,
 );
 
-#[derive(Debug, Serialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
 #[cfg_attr(not(debug_assertions), derive(Deserialize))]
 #[serde(transparent)]
 pub struct FileTree {
     pub node: BTreeMap<String, FileTreeEntry>,
 }
 
-impl FileTree {
-    pub fn iter_dfs<'iter>(&'iter self) -> FileTreeDFS<'iter> {
+impl<'iter> FileTree {
+    pub fn iter_dfs(&'iter self) -> FileTreeDepthFirstIter<'iter> {
         //let mut iters = VecDeque::new();
         //iters.push_front(self.node.iter());
-        let iters: VecDeque<_> = [(PathBuf::from("./"), self.node.iter())].into();
+        let iters: VecDeque<_> = [(vec!["./"], self.node.iter())].into();
 
-       FileTreeDFS {
-            current_dir: iters.front().unwrap().0.as_path(),
+        FileTreeDepthFirstIter {
+            tree: PhantomData,
             iters,
         }
     }
@@ -140,33 +144,166 @@ impl<'de> Deserialize<'de> for FileTree {
     }
 }
 
-#[derive(Debug)]
-pub struct FileTreePathView {
-    pub directory: PathBuf,
-    pub name: String,
+/// A view of a file yielded by a tree iterator.
+///
+/// Paths are represented as invididual components stored in a vector.
+/// The path: ./alienwarpowers/models/dumbbert.mdl
+/// Would be represented as:
+///
+/// ```rust
+/// use crate::{FileTreePathView, FileTreeInfo};
+///
+/// let dumbbert = FileTreePathView {
+///        directory: vec!["./", "alienwarpowers", "models"],
+///        name: "dumbbert.mdl",
+///        file_info: &FileTreeInfo {
+///            attr: None,
+///            length: 1.try_into().unwrap(),
+///            pieces_root: None,
+///        },
+///    };
+/// ```
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct FileTreePathView<'iter> {
+    /// Directory path components.
+    pub directory: Vec<&'iter str>,
+    /// File name.
+    pub name: &'iter str,
+    /// Length and hashes for the file.
+    pub file_info: &'iter FileTreeInfo,
 }
 
-pub struct FileTreeDFS<'iter> {
-    current_dir: &'iter Path,
-    iters: VecDeque<(PathBuf, btree_map::Iter<'iter, String, FileTreeEntry>)>,
+/// Depth first iterator for [FileTree].
+pub struct FileTreeDepthFirstIter<'iter> {
+    // The iterator returns references to strings held by an instance of FileTree, but it doesn't need to own it.
+    tree: PhantomData<&'iter FileTree>,
+    // Holds iterators produced by traversing the FileTree as well as keeps directory state (see implementation).
+    iters: VecDeque<(
+        Vec<&'iter str>,
+        btree_map::Iter<'iter, String, FileTreeEntry>,
+    )>,
 }
 
-impl<'iter> Iterator for FileTreeDFS<'iter> {
-    type Item = FileTreePathView;
+impl<'iter> Iterator for FileTreeDepthFirstIter<'iter> {
+    type Item = FileTreePathView<'iter>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        let cur_iter = self.iters.front()?;
-        // let (name, entry) = cur_iter.next()?;
+        let (directory, mut cur_iter) = self.iters.pop_front()?;
 
-        None
+        match cur_iter.next() {
+            Some((name, entry)) => match &entry.0 {
+                Either::Left(file_info) => {
+                    // I can't return a slice because it's owned by the iterator.
+                    let directory_view = directory.clone();
+                    // The iterator yielded a file therefore it needs to be checked again on the next call to...next().
+                    self.iters.push_front((directory, cur_iter));
+
+                    Some(FileTreePathView {
+                        directory: directory_view,
+                        name: name.as_str(),
+                        file_info,
+                    })
+                }
+                Either::Right(dir) => {
+                    // The iterator yielded a directory so the NEXT directory is the old directory with the next path name appended.
+                    let mut directory = directory.clone();
+                    directory.push(name.as_str());
+
+                    // As this is depth first, the next iterator is the next directory rather than exhausting the current iterator.
+                    self.iters.push_front((directory, dir.node.iter()));
+                    // Call next() to yield the next file. This is recursive and can cause a Stack Overflow with a malicious torrent.
+                    // So uh, fix it later.
+                    self.next()
+                }
+            },
+            // Current iterator has been expended; now traverse backward down the tree.
+            None => self.next(),
+        }
     }
 }
 
+impl FusedIterator for FileTreeDepthFirstIter<'_> {}
+
 #[cfg(test)]
 mod tests {
-    use super::FileTree;
+    use super::{FileTree, FileTreeEntry, FileTreeInfo, FileTreePathView};
+    use either::Either;
     use serde::{Deserialize, Serialize};
     use serde_bencode::Deserializer;
+
+    // Convenience function to return a new FileTreeEntry that's a file.
+    fn new_file<S>(name: S) -> (String, FileTreeEntry)
+    where
+        S: Into<String>,
+    {
+        (
+            name.into(),
+            FileTreeEntry(Either::Left(FileTreeInfo {
+                attr: None,
+                length: 1.try_into().unwrap(),
+                pieces_root: None,
+            })),
+        )
+    }
+
+    // Convenience function to return a new FileTreeEntry that's a directory.
+    fn new_dir(name: &str, entries: Vec<(String, FileTreeEntry)>) -> (String, FileTreeEntry) {
+        (
+            name.to_owned(),
+            FileTreeEntry(Either::Right(FileTree {
+                node: entries.into_iter().collect(),
+            })),
+        )
+    }
+
+    // A FileTree consisting of multiple files and a few nested directories.
+    fn multiple_files_tree() -> FileTree {
+        FileTree {
+            node: [
+                new_file("alienwarpowers"),
+                new_file("alienwarpowers.exe"),
+                new_dir(
+                    "assets",
+                    vec![new_dir(
+                        "audio",
+                        vec![
+                            new_dir(
+                                "music",
+                                (0..3)
+                                    .map(|n| new_file(format!("jon_music{n}.mp3")))
+                                    .collect(),
+                            ),
+                            new_dir(
+                                "sound",
+                                (0..3)
+                                    .map(|n| new_file(format!("soundlol{n}.wav")))
+                                    .collect(),
+                            ),
+                        ],
+                    )],
+                ),
+                new_dir(
+                    "media",
+                    vec![new_file("manual.pdf"), new_file("aliens.pdf")],
+                ),
+            ]
+            .into_iter()
+            .collect(),
+        }
+    }
+
+    fn pls_equal(view_one: FileTreePathView<'_>, name: &str, directory: Vec<&str>) {
+        let (file_name, file_info) = new_file(name);
+        let file_info = file_info.0.unwrap_left();
+
+        let view_two = FileTreePathView {
+            directory,
+            name: &file_name,
+            file_info: &file_info,
+        };
+
+        assert_eq!(view_one, view_two)
+    }
 
     #[derive(Deserialize, Serialize)]
     struct OuterTest {
@@ -196,5 +333,44 @@ mod tests {
         assert_eq!(BENCODE, info_se);
 
         Ok(())
+    }
+
+    #[test]
+    fn filetree_single_iter_depth() {
+        let tree = FileTree {
+            node: [new_file("raees_buggy_code.js")].into_iter().collect(),
+        };
+
+        let tree_view: Vec<_> = tree.iter_dfs().collect();
+        assert_eq!(tree_view.len(), 1);
+
+        let (file_name, file_info) = new_file("raees_buggy_code.js");
+        let file_info = file_info.0.unwrap_left();
+
+        let raees_view = FileTreePathView {
+            directory: vec!["./"],
+            name: &file_name,
+            file_info: &file_info,
+        };
+        assert_eq!(
+            tree_view
+                .into_iter()
+                .next()
+                .expect("Expected Raees' buggy code."),
+            raees_view
+        );
+    }
+
+    #[test]
+    fn filetree_multiple_iter_depth() {
+        let tree = multiple_files_tree();
+        let mut tree_view = tree.iter_dfs();
+
+        pls_equal(tree_view.next().unwrap(), "alienwarpowers", vec!["./"]);
+    }
+
+    #[test]
+    fn filetree_dirs_o_fun() {
+        //(0..100).fold()
     }
 }