+++ crates/bedit-cloudburst/src/lib.rs
@@ -10,7 +10,7 @@ mod torrent;
 mod uriwrapper;
 
 pub use fileattributes::{FileAttribute, TorrentFileAttributes};
-pub use files::{FileTree, FileTreeEntry, FileTreeInfo, SharedFiles};
+pub use files::{FileTree, FileTreeEntry, FileTreeInfo, FileTreePathView, SharedFiles};
 pub use info::{Hybrid, Info, MetaV1, MetaV2};
 pub use pieces::{PieceLength, Pieces};
 pub use torrent::{Node, Torrent};