+++ crates/bedit-cloudburst/CHANGELOG.md
@@ -2,6 +2,7 @@
 - Simplify and improve bin tester by deduplicating code to print results and iterating over `Result<Vec<PathBuf, _>>` instead of `Result<Vec<Torrent, _>>>`.
 - Add pretty colors to output.
 - Use wrappers around hash types in `Info` and everywhere else.
+- Implement a depth first iterator for `FileTree`.
 
 # 0.5.0
 - Wrap `ByteBuf` into `HexBytes`